Moved coretest::str tests into collectiontest::str
